+++ src/Service/UploaderHelper.php
@@ -2,45 +2,71 @@
 
 namespace App\Service;
 
+use App\Controller\ArticleReferenceAdminController;
 use Gedmo\Sluggable\Util\Urlizer;
 use League\Flysystem\AdapterInterface;
 use League\Flysystem\FileNotFoundException;
 use League\Flysystem\FilesystemInterface;
 use Psr\Log\LoggerInterface;
-use Symfony\Component\HttpFoundation\File\File;
 use Symfony\Component\Asset\Context\RequestStackContext;
+use Symfony\Component\HttpFoundation\File\File;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 
+/**
+ * Helper to handle the files by uploads and fixtures
+ * Class UploaderHelper
+ * @package App\Service
+ */
 class UploaderHelper
 {
     const ARTICLE_IMAGE = 'article_image';
     const ARTICLE_REFERENCE = 'article_reference';
 
-    private $filesystem;
-
-
+    //private $uploadsPath;
     private $requestStackContext;
-
+    private $publicUploadsFilesystem;
+    private $privateUploadsFilesystem;
     private $logger;
 
     private $publicAssetBaseUrl;
 
-    public function __construct(FilesystemInterface $uploadsFilesystem, RequestStackContext $requestStackContext, LoggerInterface $logger, string $uploadedAssetsBaseUrl)
+    /**
+     * Dependency Injection -we need an direction to our uploads path
+     * UploaderHelper constructor.
+     * //@param string $uploadsPath
+     * If you're using version 4 of oneup/flysystem-bundle (so, flysystem v2), autowire Filesystem instead of FilesystemInterface from League\Flysystem.
+     * @param string $uploadedAssetsBaseUrl -@see services.yaml
+     *
+     * If you're using version 4 of oneup/flysystem-bundle (so, flysystem v2), autowire Filesystem instead of FilesystemInterface from League\Flysystem.
+     * @param FilesystemInterface $publicUploadsFilesystem -@see services.yaml filesystem for public files
+     * @param FilesystemInterface $privateUploadsFilesystem  -@see services.yaml filesystem for private files
+     */
+    public function __construct(FilesystemInterface $publicUploadsFilesystem /*string $uploadsPath before fileSystem*/, FilesystemInterface $privateUploadsFilesystem, RequestStackContext $requestStackContext, LoggerInterface $logger, string $uploadedAssetsBaseUrl)
     {
-        $this->filesystem = $uploadsFilesystem;
+        // $this->uploadsPath = $uploadsPath;
+        $this->publicUploadsFilesystem = $publicUploadsFilesystem;
+        $this->privateUploadsFilesystem = $privateUploadsFilesystem;
         $this->requestStackContext = $requestStackContext;
         $this->logger = $logger;
         $this->publicAssetBaseUrl = $uploadedAssetsBaseUrl;
     }
 
+    /**
+     * Method to handle uploadingArticleImage
+     * @param File $file //it was an UploadedFile (it extend File) $uploadedFile but we had to change it with File because we use it in fixture
+     * @param null|string $existingFilename name of the file that we want to use after editing
+     * @return string the name of the file
+     */
     public function uploadArticleImage(File $file, ?string $existingFilename): string
     {
+
+        // now we use public filesystem
         $newFilename = $this->uploadFile($file, self::ARTICLE_IMAGE, true);
 
+        // remove the old file
         if ($existingFilename) {
             try {
-                $result = $this->filesystem->delete(self::ARTICLE_IMAGE.'/'.$existingFilename);
-
+                $result = $this->publicUploadsFilesystem->delete(self::ARTICLE_IMAGE.'/'.$existingFilename);
                 if ($result === false) {
                     throw new \Exception(sprintf('Could not delete old uploaded file "%s"', $existingFilename));
                 }
@@ -52,6 +78,12 @@ class UploaderHelper
         return $newFilename;
     }
 
+    /**
+     * After uploadArticleReference we want put file to the privateDirectory @see ArticleReferenceAdminController
+     * @param File $file
+     * @return string
+     * @throws \League\Flysystem\FileExistsException
+     */
     public function uploadArticleReference(File $file): string
     {
         return $this->uploadFile($file, self::ARTICLE_REFERENCE, false);
@@ -70,31 +102,52 @@ class UploaderHelper
             ->getBasePath().$fullPath;
     }
 
+
     /**
+     * Get Resource of file from path depending on the isPublic variable
+     * @see ArticleReferenceAdminController::downloadArticleReference()
      * @return resource
      */
-    public function readStream(string $path)
+    public function readStream(string $path, bool $isPublic)
     {
-        $resource = $this->filesystem->readStream($path);
-
+        $filesystem = $isPublic ? $this->publicUploadsFilesystem : $this->privateUploadsFilesystem;
+        $resource = $filesystem->readStream($path);
         if ($resource === false) {
             throw new \Exception(sprintf('Error opening stream for "%s"', $path));
         }
-
         return $resource;
     }
 
-    public function deleteFile(string $path)
+    /**
+     * Handle deleting the file
+     * @param string $path
+     * @param bool $isPublic
+     * @throws \Exception
+     */
+    public function deleteFile(string $path, bool $isPublic)
     {
-        $result = $this->filesystem->delete($path);
-
+        $filesystem = $isPublic ? $this->publicUploadsFilesystem : $this->privateUploadsFilesystem;
+        $result = $filesystem->delete($path);
         if ($result === false) {
             throw new \Exception(sprintf('Error deleting "%s"', $path));
         }
     }
 
-    private function uploadFile(File $file, string $directory, bool $isPublic): string
+    /**
+     * Handle the file by private filesystem and public filesystem
+     * @param File $file
+     * @param string $directory where to save: @see uploadArticleReference (private) and @see uploadArticleImage (public)
+     * @param bool $isPublic which filesystem we want to use
+     * @return string
+     * @throws \League\Flysystem\FileExistsException
+     */
+    private function uploadFile(File $file, string $directory, bool $isPublic)
     {
+        /* Before filesystem
+            $destination = $this->uploadsPath.'/'.self::ARTICLE_IMAGE;
+         */
+
+        //if we upload we have an UploadedFile, but in fixtures we use File
         if ($file instanceof UploadedFile) {
             $originalFilename = $file->getClientOriginalName();
         } else {
@@ -102,8 +155,28 @@ class UploaderHelper
         }
         $newFilename = Urlizer::urlize(pathinfo($originalFilename, PATHINFO_FILENAME)).'-'.uniqid().'.'.$file->guessExtension();
 
+        /* Before filesystem
+        $file->move(
+            $destination,
+            $newFilename
+        );
+        */
+
+        //create new file with content
+        /*
+        $this->publicUploadsFilesystem->write( //this (file_get_contents) reads the entire contents of the file into PHP's memory.
+            self::ARTICLE_IMAGE.'/'.$newFilename,
+            file_get_contents($file->getPathname())
+        );
+        */
+
+        // we want to use specific filesystem in dependence if it is public or not
+        $filesystem = $isPublic ? $this->publicUploadsFilesystem : $this->privateUploadsFilesystem;
+
+        //create new file with content in less memory using way
         $stream = fopen($file->getPathname(), 'r');
-        $result = $this->filesystem->writeStream(
+        $result = $filesystem->writeStream(
+        // before we use self::ARTICLE_IMAGE.'/'.$newFilename, but now we handle public and private types
             $directory.'/'.$newFilename,
             $stream,
             [
@@ -115,10 +188,10 @@ class UploaderHelper
             throw new \Exception(sprintf('Could not write uploaded file "%s"', $newFilename));
         }
 
+        // we want to close stream after writing
         if (is_resource($stream)) {
             fclose($stream);
         }
-
         return $newFilename;
     }
 }