+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Final
+## [05.04.2023]
+- Final merging Mailer and WepackEncore and Filesystem projects
 # Symfony Mailer: Love Sending Emails Again
 ## [20.02.2022]
 - Processing Encore Files through inline_css()
@@ -37,3 +40,68 @@
 - Transport Config & Mailtrap
 - Creating, Configuring & Sending the Email Object
 - Hello Symfony Mailer
+# Webpack Encore: Frontend like a Pro!
+MERGE
+# Webpack Encore: Frontend like a Pro!
+## [02.04.2022]
+- Production Build & Deployment
+- Async Imports
+- The Single Runtime Chunk
+- Polyfills & Babel
+- browserslist: What Browsers do you need to Support?
+- Support any Browser with PostCSS & Babel
+- addStyleEntry(): CSS-Only Entrypoint
+- Auto-Provide jQuery for Mischievous Packages
+- Refactoring Autocomplete JS & CSS
+- Entry Refactoring
+- Page-Specific JS: Multiple Entrie
+## [19.03.2022]
+- Copying Files
+- Sass & Overriding Bootstrap Variables
+- Importing 3rd Party CSS + Image Paths
+- The Magic of Split Chunks
+- Bootstrap & the Curious Case of jQuery Plugins
+- Importing External Libraries & Global Variables
+- Modules: require() & import()
+- Twig Helpers, entrypoints.json & yarn Scripts
+- Webpacking our First Assets
+- Hello Webpack Encore
+# All about Uploading Files in Symfony
+## [12.03.2022]
+- Reordering Endpoint & AJAX
+- Reordering the Files
+## [04.03.2022]
+- JavaScr ipt for Editing a Reference
+- Edit Endpoint & Deserialization
+- Deleting Files
+- Rendering the File List Client Side
+- API Endpoint & Errors with Dropzone
+- Dropzone: AJAX Upload
+- Streaming the File Download
+- Endpoint for Downloading Private Files
+- Mime Type Validation
+## [04.03.2022]
+- Storing Private Files
+- Uploading References
+- Setup for Uploading Private Article References
+- Absolute Asset Paths
+## [01.03.2022]
+- Flysystem <3 LiipImagineBundle
+- Flysystem: Streaming & Defensive Coding
+- Using the Filesystem
+- Flysystem: Filesystem Abstraction
+## [27.02.2022]
+- File Uploads & Data Fixtures
+- Image Preview on the Form
+- Thumbnailing with LiipImagineBundle
+## [26.02.2022]
+- The asset() Function & assets.context
+- URL to Public Assets
+- Upload Field Styling & Bootstrap
+- File Validation
+- Centralizing Upload Logic
+- File Upload Field in a Form
+- Unique (but not Insane) Filenames
+- Where & How to Store the File
+- Uploads, multipart/form-data & UploadedFile
+- Setting up with the Symfony Local Web Server